fix: add unban-all-peers
